fix query.rs
