+++ crates/bvs-strategy-base/src/lib.rs
@@ -4,7 +4,6 @@ pub mod state;
 
 mod auth;
 pub mod error;
-pub mod query;
 
 pub use crate::error::ContractError;
