+++ crates/bvs-strategy-base/src/msg.rs
@@ -1,11 +1,6 @@
-use crate::query::{
-    SharesResponse, SharesToUnderlyingResponse, StrategyManagerResponse, TotalSharesResponse,
-    UnderlyingResponse, UnderlyingToShareResponse, UnderlyingToSharesResponse,
-    UnderlyingTokenResponse,
-};
 use crate::state::StrategyState;
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::Uint128;
+use cosmwasm_std::{Addr, Uint128};
 
 #[cw_serde]
 pub struct InstantiateMsg {
@@ -57,6 +52,41 @@ pub enum QueryMsg {
 
     #[returns(StrategyState)]
     GetStrategyState {},
+}
+
+#[cw_serde]
+pub struct SharesResponse {
+    pub total_shares: Uint128,
+}
+
+#[cw_serde]
+pub struct UnderlyingResponse {
+    pub amount_to_send: Uint128,
+}
+
+#[cw_serde]
+pub struct SharesToUnderlyingResponse {
+    pub amount_to_send: Uint128,
+}
+
+#[cw_serde]
+pub struct UnderlyingToSharesResponse {
+    pub share_to_send: Uint128,
+}
+
+#[cw_serde]
+pub struct StrategyManagerResponse {
+    pub strategy_manager_addr: Addr,
+}
+
+#[cw_serde]
+pub struct UnderlyingTokenResponse {
+    pub underlying_token_addr: Addr,
+}
+
+#[cw_serde]
+pub struct TotalSharesResponse {
+    pub total_shares: Uint128,
 }
 
 /// Both Strategy Base & Strategy Manager circularly depend on each other.
